refactor(web): formalize frontend file budgets
Summary:
- replace per-file frontend budget waivers with named budget categories
  shared by the structure script, ESLint config, and README docs
- split the oversized board and project-conversation regression suites
  into focused files with shared helpers so they fit the new policy
- update the skills editor Playwright flow and related tests to follow
  the CodeMirror-backed editor surface used in the app today

Rationale:
- the old allowlist and `max-lines` bypasses were temporary debt that
  made the file-budget rules drift between tools and docs
- promoting recurring file shapes into first-class categories keeps the
  policy reviewable while still forcing true outliers to split
- keeping the e2e selector aligned with the actual editor prevents the
  workflow-aligned frontend gate from failing on a stale textarea probe

Tests:
- `PATH=$HOME/.nvm/versions/node/v22.22.1/bin:$PATH corepack pnpm run lint`
- `PATH=$HOME/.nvm/versions/node/v22.22.1/bin:$PATH corepack pnpm run lint:structure`
- `PATH=$HOME/.nvm/versions/node/v22.22.1/bin:$PATH corepack pnpm run lint:deps`
- `PATH=$HOME/.nvm/versions/node/v22.22.1/bin:$PATH corepack pnpm run check`
- `PATH=$HOME/.nvm/versions/node/v22.22.1/bin:$PATH corepack pnpm exec vitest run --reporter=dot src/lib/features/board/components/board-page-cache-and-ticket-controls.test.ts src/lib/features/board/components/board-page-column-actions.test.ts src/lib/features/board/components/board-page-archive-actions.test.ts src/lib/features/chat/project-conversation-controller-turns.test.ts src/lib/features/chat/project-conversation-controller-tabs.test.ts src/lib/features/chat/project-conversation-workspace-browser-navigation.test.ts src/lib/features/chat/project-conversation-workspace-browser-refresh.test.ts src/lib/features/chat/project-conversation-controller-restore.test.ts src/lib/features/chat/project-conversation-event-bus.test.ts src/lib/features/board/components/board-page-runtime-and-drawer.test.ts src/lib/features/chat/project-conversation-panel-session.test.ts src/lib/features/settings/components/access-settings.test.ts`
- `PATH=$HOME/.nvm/versions/node/v22.22.1/bin:$PATH corepack pnpm exec playwright test tests/e2e/skills-editor.spec.ts --reporter=dot --quiet`
- `PATH=$HOME/.nvm/versions/node/v22.22.1/bin:$PATH corepack pnpm run build`
- `PATH=$HOME/.nvm/versions/node/v22.22.1/bin:$PATH corepack pnpm run ci`
