+++ web/src/lib/features/chat/project-conversation-controller-restore.test.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-lines */
 import { afterEach, describe, expect, it, vi } from 'vitest'
 
 const {
