+++ web/src/lib/features/chat/project-conversation-event-bus.test.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-lines */
 import { afterEach, describe, expect, it, vi } from 'vitest'
 
 const { watchProjectConversationMuxStream } = vi.hoisted(() => ({
