+++ web/src/lib/features/chat/project-conversation-workspace-browser-state.svelte.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-lines */
 import {
   getProjectConversationWorkspace,
   getProjectConversationWorkspaceFilePatch,
