+++ web/src/lib/features/chat/project-conversation-workspace-browser.test-helpers.ts
@@ -1,0 +1,68 @@
+import type { ProjectConversationWorkspaceDiff } from '$lib/api/chat'
+
+export const workspaceMetadata = {
+  conversationId: 'conversation-1',
+  available: true,
+  workspacePath: '/tmp/conversation-1',
+  repos: [
+    {
+      name: 'openase',
+      path: 'services/openase',
+      branch: 'agent/conv-123',
+      headCommit: '123456789abc',
+      headSummary: 'Add workspace browser scaffolding',
+      dirty: true,
+      filesChanged: 1,
+      added: 2,
+      removed: 0,
+    },
+  ],
+}
+
+export const workspaceDiff = {
+  conversationId: 'conversation-1',
+  workspacePath: '/tmp/conversation-1',
+  dirty: true,
+  reposChanged: 1,
+  filesChanged: 1,
+  added: 2,
+  removed: 0,
+  repos: [
+    {
+      name: 'openase',
+      path: 'services/openase',
+      branch: 'agent/conv-123',
+      dirty: true,
+      filesChanged: 1,
+      added: 2,
+      removed: 0,
+      files: [{ path: 'README.md', status: 'modified', added: 2, removed: 0 }],
+    },
+  ],
+} satisfies ProjectConversationWorkspaceDiff
+
+export function mockWorkspaceMetadata(getProjectConversationWorkspace: {
+  mockResolvedValue: (value: unknown) => void
+}) {
+  getProjectConversationWorkspace.mockResolvedValue({
+    workspace: structuredClone(workspaceMetadata),
+  })
+}
+
+export function deferredPromise<T>() {
+  let resolve!: (value: T | PromiseLike<T>) => void
+  let reject!: (reason?: unknown) => void
+  const promise = new Promise<T>((nextResolve, nextReject) => {
+    resolve = nextResolve
+    reject = nextReject
+  })
+  return { promise, resolve, reject }
+}
+
+export function ensureResizeObserver() {
+  globalThis.ResizeObserver ??= class {
+    observe() {}
+    unobserve() {}
+    disconnect() {}
+  }
+}
