+++ web/src/lib/features/settings/components/security-settings-secret-bindings.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-  /* eslint-disable max-lines */
   import type { ScopedSecret, ScopedSecretBinding, Ticket, Workflow } from '$lib/api/contracts'
   import { Badge } from '$ui/badge'
   import { Button } from '$ui/button'
