+++ web/src/lib/features/skills/components/skill-editor-page.test.ts
@@ -1,5 +1,3 @@
-/* eslint-disable max-lines */
-
 import { cleanup, fireEvent, render, waitFor } from '@testing-library/svelte'
 import { afterEach, beforeAll, describe, expect, it, vi } from 'vitest'
